:sparkles: feat: Create get items from DB

+++ app/src/main/java/com/app/ifplan_leite/core/data/state/SimulateItems.kt
@@ -4,10 +4,9 @@ import androidx.room.TypeConverters
 import com.app.ifplan_leite.core.data.model.utils.IFPlanConverters
 import kotlinx.serialization.Serializable
 import java.sql.Date
-
 @Serializable
 data class SimulateItems(
-    var id: Int,
+    var id: Long,
     var title: String,
     var creationDate: String,
     var description: String
